get rid of Component
